Add files via upload
Ultimately, it was too difficult to commit individual changes to each file. I have uploaded all of the updated files. I have added a few classes for the leaderboard and scoring system. In addition, I have improved the functionality of some aspects of the game, fixing some of the bugs. I have also added background music for the game and other small features for a smooth and streamlined experience